Rename usernameSlice to UserSlice
