+++ src/components/username/userSlice.ts
@@ -3,18 +3,18 @@ import type { AppState } from "@store";
 import { LogEntry } from "models/Calendar";
 import { User } from "models/User";
 
-export interface UsernameState {
+export interface UserState {
   value?: User;
   byDay?: LogEntry;
   status: "idle" | "loading" | "failed";
 }
 
-const initialState: UsernameState = {
+const initialState: UserState = {
   value: undefined,
   status: "idle",
 };
 
-export const usernameSlice = createSlice({
+export const userSlice = createSlice({
   name: "username",
   initialState,
   reducers: {
@@ -24,11 +24,11 @@ export const usernameSlice = createSlice({
   },
 });
 
-export const { setUsername } = usernameSlice.actions;
+export const { setUsername } = userSlice.actions;
 
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.username.value)`
 export const selectUsername = (state: AppState) => state.username.value;
 
-export const usernameReducer = usernameSlice.reducer;
+export const usernameReducer = userSlice.reducer;
